+++ BlockTelemetry/mod_info.js
@@ -5,7 +5,7 @@
    Name : "Block Telemetry",
    Description : "Force disable telemetry and block any telemetry event that ignore the flags",
    Url: {
-      "Nexus"  : "https://nexusmods.com/phoenixpoint/mods/45/",
+      "Nexus"  : "https://nexusmods.com/phoenixpoint/mods/48/",
       "GitHub" : "https://github.com/Sheep-y/PhoenixPt-Mods/"
    },
 })