+++ BlockTelemetry/changelog.md
@@ -1,5 +1,7 @@
 Changelog of Block Telemetry, a Phoenix Point Mod by Sheepy
 
-# Version 1, 2020-05-02
+# Version 1, 2020-05-03
 
-* Force disable telemetry and block telemetry events.
+* Force disable telemetry.
+* Block telemetry events.
+* Disable telemetry option checkbox.